agents-v2/requirements.txt
# Guardian AI v2 - Google ADK Multi-Agent System
# Python 3.10+

# Core dependencies
pydantic>=2.0.0
networkx>=3.0

# Google ADK - Agent Development Kit (MANDATORY)
google-adk>=1.0.0

# Google AI Platform
google-generativeai>=0.3.0

# Optional: For web interface
fastapi>=0.100.0
uvicorn>=0.23.0

# Development tools
pytest>=7.4.0
black>=23.0.0